all commands successfully implemented
