+++ src/main/java/DialogBox.java
@@ -1,5 +1,3 @@
-package duke;
-
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.geometry.Pos;
